Fix problems with redraws when resting
Really hacky "solution". We listen to EVENT_ANIMATE, and, if the
player is not resting, animate monsters and redraw the screen
(this also animates player running).
We also listen to EVENT_END and always redraw the screen on it.
redraw_stuff() have some code to make resting (and running) go faster.
All of it looks pretty bad, hopefully a better solution can be found...