Refactor key_confirm_command and remove "^*"
The inscription "^*" prevented the player from moving (in old text
ui, too). That's because movement is two commands (';', dir) and
the second one was handled by textui_get_check() as "no" (that is,
"dont move"). The players don't seem to complain about it, presumably
because no one uses "^*". Therefore, it was removed.