Change the logic of display redraws
(as discussed on oook) there are essentially only two cases where
display redraws must be performed:
a) When the user can take an action (as in, press a key)
b) When some kind of animation is played
inkey_full() already takes care of the first case. It follows that
ui2-display.c should only concern itself with animations.
Displaying player running is a sort of animation (we don't want @
to teleport from one end of corridor to the other). Fortunately,
the core sends appropriate event for that: EVENT_PLAYERMOVED. We
redraw the display when it is recieved.
Another sort of animation is display of resting counter (the event
in this case is EVENT_STATE). We don't want to its show every tick,
though, because that would take too long (some users like to scum
for things in shops etc). So we implement frame-skipping logic,
similar to that in redraw_stuff() (except we don't call it a "hack").
This is a lot more effecient and reasonable solution than what I came
up with previously.
Other such cases (of when to show stuff) should be dealt with similarly.