Add member "has_flags" to term points
As an optimization. "has_flags" does what it says; if a point has
some flags, it must be true, else it must be false. That way the
term code won't have to call tpf_is_equal() on points, unless it's
necessary, and, more importantly, the frontend code wont have to
call tpf_has() on every single point.
(when/if some flags get added, that is :)