Fix map redraws when player is in the darkness (with no light)
cave_unlight() sets FEAT_NONE on all darkened floor squares (called
a "hack" in comment), while print_map() uses FEAT_NONE to decide if
a square has to be drawn or left blank (also a hack?). That leads to
player tile not being drawn, if '@' finished running on a darkened
square, because disturb() asks for whole-map redraw, which is a hack
that serves as a workaround for a hack in redraw_stuff().
Fun, isn't it?