Optimize map redraws
Another optimization. It doesn't actually have much effect, since
map redraws are not a bottleneck. The main value is that it can be
extended to frontends (if necessary) and even used to implement
smooth scrolling of the map (unfortunately, Angband's monsters
dont move smoothly, they "teleport" instead!).
This is quite experimental and can get reverted.