Optimize map redraws with SDL2
This is a big one. With "center_player" on, it allows to reduce the
number of calls to SDL_RenderCopy from thousands to two (SDL2 is a
very nice library, but it does have significant drawbacks, such as
not batching any output - so we do it manually).
Anyway, any frontend can make use of it, and with this optimization,
it is incorrect to say that "center_player" is slow. It is about as
fast as any other redraw by any amount of grids (at least with SDL2).
It doesn't work with double tiles (so far), so it is just disabled
when they're used ("old-style" redraws are still possible, of course).
Double tiles cause all sorts of problems :)
It it still experimental, but unlikely to get reverted.