Refactor double tiles
To make it work with new optimizations, we just track the number
of double-tiled thing on screen. If there are any, we fall back to
the old way of redrawing the map; otherwise, use the new way. This
allows to keep the old way of handling double tiles, which is nice
and relatively simple.
The alternative is to copy information about double tiles and adjust
it to new coordinates... maybe later?
For now, I've had enough of stone giants :)