Increate key repeat rate
Just to test scrolling (of various things). For instance, SDL2
on my machine chockes in options menu ("User interface options")
because it has a lot of grids. Anyway, the only time it's a good
idea to hold down a key in Angband is when reading a help file,
and who reads those? :)