Optimize messages reader
Same as with the help reader.
That must be it for optimizing scrolling. The only other things
that scrolls is textblocks, but in practice, there are no textblocks
that are big enough to make scrolling necessary. And making players
read such walls of text is probably not a good idea anyway...