Simplify term handling in frontends
There is no reason actually to have pointer to term in a frontend.
About the only thing that can be done with it is to resize the term;
temporary terms should not be resized (at least, not by a frontend)
and resizing permanent terms can be handled by ui2-display.c.